Escape the ID to prevent an uncaught syntax error when a page === heading === ends in a dot
For example, if a (NS_MAIN) page contained the header === Star Wars etc. ===, the previous version would cause this to be logged in the browser's console:
Uncaught Error: Syntax error, unrecognized expression: #Star_Wars_etc.

"Star_Wars_etc." is a valid HTML ID, but when used here, the dots need to be \\escaped.
jQuery's $.escapeSelector() conveniently does this for us.

Thanks to MatmaRex for the help with this one!